pyproject.toml
[tool.poetry]
name = "gerador-conteudo"
version = "1.0.0"
description = "Estrutura para desenvolvimento de POCs em Python"
authors = ["Seu Nome <seu.email@exemplo.com>"]
readme = "README.md"
packages = [{include = "pocs"}, {include = "scripts"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.14"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.group.data.dependencies]
pandas = "^2.0.0"

[tool.poetry.group.api.dependencies]
fastapi = "^0.100.0"
uvicorn = "^0.23.0"

[tool.poetry.group.automation.dependencies]
selenium = "^4.15.0"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.social.dependencies]
# Dependências para APIs de redes sociais
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.ai.dependencies]
# Dependências para IA e geração de conteúdo
google-generativeai = "^0.3.0"
pillow = "^10.0.0"
opencv-python = "^4.8.0"
moviepy = "^1.0.3"

[tool.poetry.group.web.dependencies]
# Dependências para interface web
streamlit = "^1.28.0"
plotly = "^5.17.0"
boto3 = "^1.34.0"

[tool.poetry.group.database.dependencies]
# Dependências para banco de dados
sqlalchemy = "^2.0.0"

[tool.poetry.scripts]
run-poc = "scripts.run_poc:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
]
